'first_commit'

+++ face_reco_from_camera_ot_single_person.py
@@ -345,8 +345,8 @@ class Face_Recognizer:
     
 
                 #self.update_fps()
-                w=(screenwidth-100)/2
-                h=(screenheight-100)/2
+                w=int((screenwidth-100)/2)
+                h=int((screenheight-100)/2)
                 cv2.namedWindow("camera", 1)
                 cv2.moveWindow('camera',w,h)
                 cv2.imshow("camera", img_rd)
